ec2 backup

+++ utils/s3.py
@@ -30,15 +30,22 @@ def upload_file_to_s3(filename, file_path, user_id, project_id):
 
 
 def delete_file_from_s3(user_id, project_id):
-    s3_path = final_destination + user_id + "/" + project_id + "/Training Document/"
+    prefix_to_delete = final_destination + user_id + "/" + project_id 
     s3 = boto3.client(
         "s3",
         aws_access_key_id=AWS_ACCESS_KEY,
         aws_secret_access_key=AWS_SECRET_KEY,
         region_name='us-east-1'
     )
+
     try:
-        s3.delete_object(Bucket=bucket_name, Key=s3_path)
-        print("File deleted successfully from S3:", s3_path)
+        response = s3.list_objects_v2(Bucket=bucket_name, Prefix=prefix_to_delete)
+        if "Contents" in response:
+            for obj in response["Contents"]:
+                s3_path = obj["Key"]
+                print("Deleting:", s3_path)
+                s3.delete_object(Bucket=bucket_name, Key=s3_path)
+
+        print("Objects under prefix deleted successfully:", prefix_to_delete)
     except ClientError as e:
-        print("Error deleting file from S3:", str(e))
+        print("Error deleting objects:", str(e))
